Update Go version in GitHub workflows
Updated the Go version in GitHub workflows from 1.18 and 1.19 to 1.21 for better compatibility with newer features and performance improvements. The golangci-lint version was also updated for more recent bug-fixes and changes. These updates ensure that the CI environment used for automated testing and linting is in sync with the latest Go language version, hence provides better support for the latest Go language syntax, features, and modules.